validacoes de update user para admin

+++ src/main/java/jonatasSantos/royalLux/core/application/usecases/user/UserDeleteUseCaseImpl.java
@@ -3,7 +3,9 @@ package jonatasSantos.royalLux.core.application.usecases.user;
 import jakarta.persistence.EntityNotFoundException;
 import jonatasSantos.royalLux.core.application.contracts.repositories.UserRepository;
 import jonatasSantos.royalLux.core.application.contracts.usecases.user.UserDeleteUseCase;
+import jonatasSantos.royalLux.core.application.exceptions.UnauthorizedException;
 import jonatasSantos.royalLux.core.application.models.dtos.user.UserDeleteUseCaseOutputDto;
+import jonatasSantos.royalLux.core.domain.valueobjects.UserRole;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -18,6 +20,9 @@ public class UserDeleteUseCaseImpl implements UserDeleteUseCase {
     @Override
     public UserDeleteUseCaseOutputDto execute(Integer id) {
         var user = this.userRepository.findById(id.toString()).orElseThrow(() -> new EntityNotFoundException("Usuário inexistente"));
+
+        if(user.getRole().equals(UserRole.ADMIN))
+            throw new UnauthorizedException("Admin não pode ser deletado");
 
         this.userRepository.delete(user);
